tools/eupak: Heavily refactor tasks.
This starts the beginning of probably one of the bigger refactor jobs, and one I've been hinting at and FIXME jabbing for a while now.

Basically, tasks now are a lot closer coupled to their arguments. This is a good thing since it means that changing or adding features only requires modifying the task, not needlessly modifying main over and over.

CreateTask is the only one that works currently for this commit and has been converted to the new ITask model.