doc: Document eupak command line
This adds documentation on eupak's command line usage, which should be helpful.

This is written in a quasi-man format; the implication is that on POSIX systems we can provide this when installing binaries. Pandoc or something will be useful for doing that if I choose to. If I decide not to bother it can be rewritten anyhow.